perf(style): 静的ヘッダを即時描画しデータ部分をSuspenseでストリーミング
/style 遷移時に静的なタイトル/説明までスケルトンで覆われていた問題を解消。
ページ先頭の await(connection / 認証 / プリセット / プロフィール / 累計数)を
やめ、静的ヘッダを即描画。ユーザー依存部分を分離してストリーミングする。

- StylePageBody: 認証・プリセット・残高・生成結果を隔離し <Suspense> 配下に。
  connection() もこの動的領域内へ移動
- StyleTotalGenerationCount: 累計生成数バナーを独立 Suspense 化
  (高さ確保スケルトンでレイアウトシフト防止)
- 本番ビルドで /style は ◐ 部分プリレンダリング維持。低速回線でも
  タイトルが先に実テキスト表示され、本文だけスケルトン→差し替えを確認
